src/products/src/requirements.txt
annotated-types==0.6.0
anthropic-bedrock==0.8.0
anyio==3.7.1
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-instrumentation-flask==0.51b0
opentelemetry-instrumentation-requests==0.51b0
opentelemetry-exporter-otlp==1.30.0
blinker==1.9.0
boto3==1.35.71
botocore==1.35.71
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
distro==1.8.0
filelock==3.13.1
Flask==3.1.0
Flask-Cors==5.0.0
fsspec==2023.10.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
packaging==23.2
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
python-dateutil==2.8.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.4
six==1.16.0
sniffio==1.3.0
tokenizers==0.15.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==1.26.18
uuid==1.30
Werkzeug==3.1.3
wrapt==1.16.0
